+++ src/test-utils.js
@@ -1,0 +1,3 @@
+// This is not the same as `global.ArrayBuffer`. See:
+// https://github.com/facebook/jest/issues/7780
+export const NativeArrayBuffer = new Uint8Array().buffer.constructor;